Document assumptions for cleaning part
This adds a new section explicitly documenting some of the key 
assumptions being made in the cleaning script.

It also updates the suggestions for how people might deal with the bit 
about "looping" over all the files and deleting the ones that contain 
"DELETE ME!" so that the suggestions are more in line with best 
practices for `bash` scripting.