Add a new test that specifically checks internal directory structure
Students often had problems where they had all the correct files, but got the "internal directory structure" of the new archive wrong. E.g., instead of having `little_dir/file_1` they'd have `./file_1`. This checks for that specifically as a separate test. There are other tests that tested for it indirectly, but if they failed it wasn't obvious that this was the problem.