Remove the temp directory argument from cleaning exercise
This removes the second "temporary directory" argument from the script tests. The tests still need to make sure that the script doesn't leave those directories behind, but I'll deal with that later.

I also added checks to `setup()` that confirms that the script file and the archive files exist.

I removed one test that assumed we had access to the expanded contents of one of the archives. Since that's not true anymore, that test really can't happen.